Create LetsEncryptCertificateRenewal.ps1
In order to run this on a schedule you will need to install-module posh-acme
You'll also need to register a certificate the first time.
I don't have a guide on this, but if you attempt to use this as a first step you will fail.
Once you have successfully ordered your first certificate (new-paorder) then you can renew it automatically using this script.